Added ModFact_PreCompute
Computing Factorial is of O(n) time complexity. So, if we are asked to calculate for q queries, the complexity becomes O(n*q) which can easily give tle if q and n are large. So, we just precompute the factorials. This changes the time complexity to O(q+n), which can run smoothly.